switch to vite support
