+++ demo-v2.7/demo.html
@@ -3,7 +3,7 @@
 <head>
   <meta charset="UTF-8">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>Vue 2 with Vite</title>
+  <title>Vue 2.7 with Vite</title>
 </head>
 <body>
   <div id="app"></div>
