+++ demo-v2.7/index.ts
@@ -1,7 +1,11 @@
 import Vue from 'vue';
 import App from './App.vue';
-
 new Vue({
   template: '<App/>',
-  components: {App}
+  components: { App },
+  provide() {
+    return {
+      sampleMessage: 'hello',
+    };
+  },
 }).$mount('#app');
