+++ demo/Cell.vue
@@ -1,29 +1,28 @@
 <template>
-  <button @click="clickTest">Test #{{ count }}</button>
+  <button @click="clickTest">Test {{count + ' Custom Props: ' + temp}}</button>
 </template>
 <script lang="ts">
-import Vue, { PropType } from 'vue';
+interface Model {
+  count: number;
+}
+
+import Vue, { PropType } from "vue";
 export default Vue.extend({
   props: {
     model: {
-      type: Object as PropType<{
-        count: number;
-      }>,
+      type: Object as PropType<Model>
     },
-    customPropSample: Number,
+    temp: Number
   },
-  inject: ['sampleMessage'],
   computed: {
     count(): number {
       return this.model.count || 0;
-    },
+    }
   },
   methods: {
-    clickTest() {
-      // @ts-ignore
-      console.log('test', this.customPropSample, this.sampleMessage);
+    clickTest(): void {
       Vue.set(this.model, 'count', this.count + 1);
-    },
-  },
+    }
+  }
 });
 </script>
