+++ demo/index.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>Vue 2.6 with Vite</title>
+</head>
+<body>
+  <div id="app"></div>
+  <script type="module" src="./index.ts"></script>
+</body>
+</html>
