+++ demo/index.ts
@@ -1,11 +1,7 @@
 import Vue from 'vue';
 import App from './App.vue';
+
 new Vue({
   template: '<App/>',
-  components: { App },
-  provide() {
-    return {
-      sampleMessage: 'hello',
-    };
-  },
+  components: {App}
 }).$mount('#app');
