+++ package.json
@@ -2,9 +2,8 @@
   "name": "@revolist/vue-datagrid",
   "version": "4.11.12",
   "description": "Vue 2 DataGrid Spreadsheet component with native Vue 2 render support",
-  "main": "./dist/vue2-datagrid.umd.cjs",
-  "module": "./dist/vue2-datagrid.mjs",
   "types": "./dist/index.d.ts",
+  "type": "module",
   "exports": {
     ".": {
       "import": "./dist/vue2-datagrid.mjs",
@@ -16,8 +15,8 @@
     "dist"
   ],
   "scripts": {
-    "build": "webpack --config ./.webpack/webpack.config.js",
-    "dev": "cd demo-wp/ && npx webpack serve --config ./webpack.dev.js",
+    "build": "vite build",
+    "dev": "vite",
     "release": "npm run build && npm publish --public --tag pre-release"
   },
   "keywords": [
@@ -65,20 +64,13 @@
     "@revolist/revogrid": "4.11.12"
   },
   "devDependencies": {
-    "@vue/compiler-sfc": "^3.2.37",
-    "clean-webpack-plugin": "^4.0.0",
-    "html-webpack-plugin": "^5.5.1",
+    "@stencil/core": "^4.24.0",
     "sass": "^1.79.3",
-    "sass-loader": "^16.0.2",
-    "style-loader": "^3.3.2",
-    "ts-loader": "^9.4.2",
     "typescript": "^5.0.4",
-    "vue": "^2.6.14",
-    "vue-loader": "^15.10.0",
-    "vue-style-loader": "^4.1.3",
-    "webpack": "^5.81.0",
-    "webpack-cli": "^5.0.2",
-    "webpack-dev-server": "^4.13.3",
-    "@stencil/core": "^4.17.1"
+    "vite": "^4.0.0",
+    "vite-plugin-dts": "^3.9.1",
+    "vite-plugin-vue2": "^2.0.3",
+    "vue-template-compiler": "^2.6.14",
+    "vue": "=2.6.14"
   }
 }
