+++ tsconfig.json
@@ -2,14 +2,15 @@
   "compilerOptions": {
     "outDir": "./dist/",
     "baseUrl": "./",
+    "esModuleInterop": true,
     "allowSyntheticDefaultImports": true,
     "allowUnreachableCode": true,
     "declaration": true,
     "experimentalDecorators": true,
-    "lib": ["dom", "es2017"],
+    "lib": ["dom", "es2017", "ESNext"],
     "moduleResolution": "node",
-    "module": "es2015",
-    "target": "es5",
+    "module": "esnext",
+    "target": "ES5",
     "strict": true,
     "noUnusedLocals": true,
     "noUnusedParameters": true,
@@ -22,5 +23,5 @@
     }
   },
   "include": ["types/jsx.d.ts", "lib", "/**/*.d.ts", "/**/*.ts", "*.d.ts"],
-  "exclude": ["node_modules", "demo", "demo-wp"]
+  "exclude": ["node_modules", "demo-v2.7", "demo"]
 }
